refactor: extract GPT helpers into separate module
