+++ src/ai.rs
@@ -1,3 +1,4 @@
 pub mod common;
+pub mod gpt;
 pub mod stt;
 pub mod vision;
