+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod ai;
 mod db;
 mod handlers;
 
+pub use ai::gpt::{parse_items_gpt, parse_voice_items_gpt};
 pub use ai::stt::{parse_items, parse_voice_items};
 pub use db::Item;
 pub use handlers::{format_delete_list, format_list, format_plain_list, parse_item_line};
